+++ CHANGELOG.md
@@ -1,3 +1,5 @@
+## UNRELEASED
+
 ## 0.4.6 (December 12, 2024)
 
 BUG FIXES: