+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 ## UNRELEASED
 
+IMPROVEMENTS:
+
+* build: Updated to Go 1.23.5 [[GH-1030](https://github.com/hashicorp/nomad-autoscaler/pull/1030)]
+
 ## 0.4.6 (December 12, 2024)
 
 BUG FIXES: