+++ CHANGELOG.md
@@ -2,12 +2,14 @@
 
 BUG FIXES:
 * plugin/strategy/threshold: do not return negative newCount
-* policy: Fix race between worker and policy hander
+* policy: Fix race between worker and policy handler
+* policy: Refactor policy handler to avoid starving the nomad server from connections
 
 IMPROVEMENTS:
 
 * build: Updated Nomad API dependency to 1.10.0 [[GH-1061](https://github.com/hashicorp/nomad-autoscaler/pull/1061)]
 * build: Updated to Go 1.24.3 [[GH-1091](https://github.com/hashicorp/nomad-autoscaler/pull/1091)]
+* plugin/APM/datadog: Add the possibility to rate limit the datadog client
 
 ## 0.4.6 (December 12, 2024)